Lots of silly little debugging.  The tutorial program now runs as far at the LCA part.  I want to get rid of the nominal focus wavelength concept.  There is a measured wavelength.  And, we can change the focus.  But we souldn't change the focus by changing the nominal focus wavelength, I don't think.  Will work on this more soon, I hope.
Also need to continue with reformatting and consistently commenting create, get, and set.  And also checking that things run correctly, as opposed to just running.  I changed enough so that there is a good chance I broke something in a calculation, probably having to do with how some size varies with wavelength.

DHB