Implemented what I think is a sensible caching procedure for the pupil function and psf.  These get computed when they don't exist or are stale, and the compute function just returns fast if they are already computed and current.
The tutorial now runs to completion again, although it is still critical to check that it is doing the right thing after all the changes.

Still hanging is how to handle the get of a diffraction limited psf, as well as the strehl ratio that depends on it.  But neither of these get functions is called in the tutorial.

Some discussion of what functions are allowed to know about the internals of the wvf structure may be warranted.  Currently, these are wvfGet, wvfSet, wvfComputePupilFunction, and wvfComputePSF.  The latter two function as the sets for the pupil function and the psf computation.  This could be folded into wvfSet, but without passing an actual argument.  But at present I think that is more confusing than the way it is now.

The tutorial is written to show the fields of this structure, which I think is a conceptual error.  Rather, there should probably be a wvfPrint or some such if we want to let the user see fields on the fly.

DHB