Small comments for this one ... but see longer notes about previous
In the previous commit, I had also updated the wvfPlot calls and tested
the validation scripts and the t_wvfZernicke script.
I made diffractionpsf get work, and strehl get seems to work.
But still, there must be errors I introduced.
On to comparing the diffraction calculations here and in ISET