Validate against early version of toolbox
Added v_wvfSVNVer121TestData script, and data that it needs.  This
shows that we can reproduce PSFs that were computed with a very early
version of the toolbox that was pretty close to what we got from Heidi.
 The one change is that we seemed to have flipped x and y relative to
the early code.  Need to figure out which is right.

This is also set up to serve as a template for making the same
comparison if we can ever get more independent test data.