Comparing Thibos and MW models
The conversion from WVF to ISET seems to be going pretty well.  In this
phase, I converted the mean Thibos data to ISET, and I compared the
calculation using the Thibos data with the MW model that is built into
ISET.  The two versions differ a bit, but not by a huge amount. The MW
model blurs the short wavelengths a little more than the Thibos model.
Otherwise, in terms of overall size and general properties, they are in
pretty good agreement.

The s_wvf2OIHuman has the comparison, all the way from a scene to the
estimated cone absorptions from the two models.