Tweaks to (mostly) s_wfv2OIHumanSamples
a) Added code that makes working directory match location of source to
several scripts.  On a mission about this so that I don't find iset
session files all over my system.

b) Explicit specification of measurement pupil size  in
s_wfv2OIHumanSamples.  This matters, because the pupil function is
computed from the Zernike coefficients taking 0-1 to be the measured
pupil size.  So with the same coefficients, you'll get a different
pupil function out depending on what you set the measurement pupil size
to.  I'm not sure what the right answer is for the Thibos model.  But
given that he gives different coefficient parameters for different
specified pupil sizes, it seems reasonable to suppose that we should
set the measured size to match the specified size.  This is what I did.
 But, it needs checking.

c) Explicit specification of measurement wavelength in
s_wvf2OIHumanSamples.  Again, I'm not sure what this should be for the
Thibos set, but setting it explicitly reminds us that we ought to
understand this.

d) Added option of computing from Hofer sample measurements, for
comparison.

e) Note that some of the changes above about explicit specification of
parameters should probably be put into other scripts that use the
Thibos data.