Cone PSF working again
The basic computation of a cone spectral sensitivity weighted PSFs is
up and running again.  Can compare with VSET version if we are clever.

Didn't yet get to optimizing focus.

Need to think through what the sce fraction means for the cone pst
case, and whether we are handling it sensibly.