Het frequency domain plots more or less working
Matlab's psf2otf leaves zero frequency in the corners or something.  I
went back to the straight fft, which I think I more or less understand,
and applied fftshift.  Now at least we get unity output at sf zero,
which we should.

Added Autrusseau data to OTF plots.  Match is in the same general
ballpark, but not perfect.  Especially for the S cones, where we fall
off much faster.  Given that this is the same Zernike coefficients
(unless there is a typo), something is not right somewhere.