Creeping closer to agreement, but not there yet.
The monochromatic PSFs we get for diffraction limited appear pretty close to what's in Figure 2 of Autresseau et al.  But the standard observer monochromatic PSFs are not so similar.  In terms of MTF, we're doing pretty well for L and M but not so well for S.

I wonder how they are computing defocus?