Added in tip and tilt to Zernike expansion.
I did this because I wondered if it might be related to why I can't
reproduce the Autrusseau numbers exactly.  It doesn't seem to have been
the problem, as adding these
just produces a wavelength independent shift of the PSF and doesn't
change the OTF.  But it seemed more rational to have them in.  Heidi's
data have these set to zero.

Also verified numerically that adding in the piston (j = 0) term
doesn't affect the PSF.  But I didn't add that in permanently, as it
would require reworking in a number of places and I don't really see
the point.