Tweaks but still don't have good agreement.
I looked at the OSA standard and am pretty sure I understand how Thibos
list the coefficients.  He does provide the j = 0 term, but we clipped
it from the table so we don't need to account for it explicitly in this
routine.  May want to make sure we handle the stochastic model
coefficients consistently, though.

My next move is to figure out how the Thibos group computes the
wavelength dependent LCA.  Maybe that isn't quite the same as the way
we are doing it.