Zeroed 4th Zernike coefficient
It occurs to me that Autresseau may have set defocus to zero at the in
focus wavelength, which we would accomplish by zeroing the 4th Z coeff.
 I did this here.  Doesn't produce exact agreement with their figures.
Not sure in the end whether they do this or not, or how to tell.

Next step remains to understand how they compute LCA.