Checking validation routines after DHB's changes.
Most of the code worked fine.  I just added so 'nowindow' or 'no fig' to
a couple of wvfPlot() calls, pulled in some missing PTB routines (into
VSET).
I started removing the cd() at the beginning of the scripts because the
new VSET doesn't save out the ISET session files.
The conversions of WVF pointspreads to VSET still seems OK.