Tidy up.  Mostly comments.  This is now in good shape.
a) Run at 1 degree pst field to match the paper.  This is actually too
small as the PSF at 400 wraps.  But it is what they did.  Also, it
doesn't matter in any significant way because the 400 nm PSF is down
weighted pretty heavily even for the S cones.

b) Verified that we reproduce 4A as well as 4B.