Fixed diffraction validation script error.
This had to do with the sign conventions on how we compute LCA, how it
is added into the pupil function, and the explicit addition of a
spectacle correction into the pupil function.  I added comments in
various places to explain this, and flipped the sign of the spectacle
correction in v_wvfDiffractionPSF to make it come out right.

I think the current conventions are rational, and match what I can find
in references.  Would be worth someone else reading the comments in the
functions listed below at some point, though, and seeing if they make
sufficient sense and are correct.