Document flip in LCA defocus sign convention
The old and new calculations have a substantive difference about the
sign of LCA.  Three possibilities:

a) This was wrong in Heidi's original code.
b) It was right in Heidi's code but we mucked it up by the time we got
to our SVN version 121.
c) It is wrong now (which I think also means that Autresseau et al.
have it wrong.)

Also note flip of y axis sign convention, and document that  as well.
I am more confident that this is right now and was wrong in Heidi's
code than I am about the LCA.