Validation script to visualize the Zernike polynomials we use
This works by getting the wavefront aberration back out when we use a
single value of 1 for various coefficients.  Useful because we can
compare what we get to pictures on the web.

The main discovery here (see comment in the validation file) is that
many of the pictures on the web are wrong.